src/requirements.txt
cloudscraper
python-dotenv
anticaptchaofficial